Test/benchmarks (#4)
We got accepted for RANLP! RANLP related experiments and data are deleted. Only important scripts are kept. RANLP-Fineweb-python is the relevant folder. For the full version checkout the previous commit.

* run_single_head.py needs more checking

* heatmaps added

* frequency code added + bug fixed in heatmap code

* topics, heads, docs visualization

* graphs for paper added

* requirements files added + extra analysis

* all used content for RANLP2025 added and refactored a bit.

* We got accepted for RANLP!
RANLP related experiments and data are deleted. Only important scripts are kept. RANLP-Fineweb-python is the relevant folder.
For the full version checkout the previous commit.